thinking_box/requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
openai-whisper>=20231117
streamlit>=1.32.0
